Take one of those and two of these Then watch the walls begin to breathe I can taste the colour of the lights Wings are growing out of me The floor is floating underneath I can see the future burning bright Future burning bright
The ceiling has us mesmerised
It feels like we could never die
Heading for the dark side of the moon
As we lift off into the sky
Invincible and so alive
Ten feet tall and fucking bulletproof
Insanity is setting in
Reality is getting thin
The universe just started slowing down
Yet suddenly we're travelling

A million miles an hour
A million miles an hour
We're going a million miles an hour
A million miles an hour

When shadows all start breaking free
No longer held by gravity
How did we get so far from the ground?
Speaking to the galaxy
Received and sending back to me
I can finally hear the speed of sound
Speed of fucking sound

I like this everlasting pill
'Cause time itself is standing still
Why's the room still spinning just the same?
Turn off your phone so no one calls
'Cause you and I are tripping balls
I can still hear ringing in my brain
Insanity is setting in
Reality is getting thin
See no signs of ever coming down
And suddenly we're travelling

A million miles an hour
A million miles an hour
We're going a million miles an hour
A million miles an hour

Insanity is setting in
Reality is getting thin
The particles are dancing all around
I can't deny the hardest thing for me
Is finding time to finally breathe
See no signs of ever coming down
Another wave is coming in

A million miles an hour
A million miles an hour
We're going a million miles an hour
A million miles an hour